+++ src/table/components/table-header.jsx
@@ -3,16 +3,14 @@ import React from "react";
 import TableCell from "@material-ui/core/TableCell";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
-import TableSortLabel from "@material-ui/core/TableSortLabel";
 import classNames from "classnames";
 import styles from "../table.module.scss";
 import { ColumnVisibilityChooser } from "./column-visibility-chooser";
-import { mapBy } from "@grailbio/lib";
+import { InnerTableHeader } from "./inner-table-header";
 
 type Props = {
   columns: Array<InternalPagedTableColumn>,
   sortingProps: SortingProps,
-  onSort?: (sortOption: SortOption) => any,
   enableSelectAll: boolean,
   hasColumnVisibilityChooser: boolean,
   columnVisibility: { [number]: boolean },
@@ -29,62 +27,30 @@ export const TableHeader = (props: Props) => {
     setColumnVisibility,
   } = props;
 
-  const { onSort, tableOptions = {} } = sortingProps;
-  if (onSort && !tableOptions.sortOptions) {
-    throw new Error("tableOptions prop is required and must include a sortOptions parameter");
-  }
-  const { sortOptions = [] } = tableOptions;
-  if (onSort && tableOptions && !sortOptions) {
-    throw new Error("tableOptions prop requires a sortOptions parameter for sorting");
-  }
-  const sortFieldsById = mapBy(sortOptions, "id");
-  const handleClickSort = (event: SyntheticMouseEvent<HTMLButtonElement>, fieldId) => {
-    const { ctrlKey } = event;
-    const currentSortField = sortFieldsById.get(fieldId) || {};
-    const newSortField = {
-      id: fieldId,
-      desc: currentSortField.id !== fieldId || currentSortField.desc !== true,
-    };
-    const sortOptions = [];
-    if (ctrlKey) {
-      // If the user holds shift while clicking on a header, multi-sort.
-      // TODO(ecarrel): support macs.
-      sortFieldsById.set(fieldId, newSortField);
-      sortOptions.push(...sortFieldsById.values());
-    } else {
-      sortOptions.push(newSortField);
-    }
-    onSort({ sortOptions });
-  };
   const visibleColumns = columns.filter(column => column.isVisible);
+  const { onSort } = sortingProps;
   return (
     <TableHead className="TableHeader">
       <TableRow>
         <>
           {visibleColumns.map((column, index) => {
             const {
-              Header, accessor = "", headerClassName = "", isSingleIcon = false, sortAccessor, id,
+              Header, accessor = "", headerClassName = "", isSingleIcon = false, sortAccessor,
             } = column;
             const fieldId = typeof accessor === "string" ? accessor : "";
             const sortFieldId = sortAccessor || fieldId;
-            const sortField = sortFieldsById.get(sortFieldId) || {};
             const isCheckboxHeader = accessor === "COLUMN_SELECT";
             const isSortable = onSort && sortFieldId.length > 0 && column.sortable !== false;
-            const isSorted = onSort && !!sortField.id;
-            const sortOrder = sortField.desc ? "desc" : "asc";
-            const key = `${id || accessor.toString() || "table-head-key"}-${index}`;
+
             let inner = Header;
             if (isSortable && !isCheckboxHeader) {
               inner = (
-                <TableSortLabel
-                  className={`sort-${sortFieldId}`}
-                  data-testid={`sort-${sortFieldId}`}
-                  active={isSorted}
-                  direction={sortOrder}
-                  onClick={event => handleClickSort(event, sortFieldId)}
+                <InnerTableHeader
+                  sortFieldId={sortAccessor || fieldId}
+                  sortingProps={sortingProps}
                 >
-                  {Header}
-                </TableSortLabel>
+                  {Header || null}
+                </InnerTableHeader>
               );
             }
             if (isCheckboxHeader && !enableSelectAll) {
@@ -92,12 +58,11 @@ export const TableHeader = (props: Props) => {
             }
             return (
               <TableCell
-                key={key}
+                key={index}
                 className={classNames(headerClassName, styles.tableHeader, {
                   [`${fieldId}-header`]: fieldId,
                   [styles.singleIcon]: isCheckboxHeader || isSingleIcon,
                 })}
-                sortDirection={isSorted ? sortOrder : false}
               >
                 {inner}
               </TableCell>