+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 40. Support `isVisible` parameter on page configs in `CommonTabbedPageV2`.
 41. Add `hasColumnVisibilityChooser` prop to `SimpleTable` and `PagedTable` which, if enabled, allows user to select
     which columns should be shown/hidden.
+42. Make LargeSimpleTable sortable.
 
 ## 1.0.0 - 2019-06-17